Implemented casting of type streamoff to size_t

+++ utils.cpp
@@ -57,7 +57,7 @@ size_t getFileLength(fstream& file) {
 	// Сдвигаем указатель на конец файла
 	file.seekg(0, ios_base::end);
 	// Узнаем, на каком месте стоит указатель
-	size_t length_in_bytes = file.tellg();
+	size_t length_in_bytes = static_cast<size_t>(file.tellg());
 	// Сдвигаем указатель назад в начало файла
 	file.seekg(0, ios_base::beg);
